+++ src/flexspi/fields.rs
@@ -99,7 +99,7 @@ pub enum SerialClockFrequency {
     MHz30 = 1,
     MHz50,
     MHz60,
-    #[cfg(not(any(feature = "imxrt1170", feature = "imxrt1180")))]
+    #[cfg(not(any(feature = "imxrt1160", feature = "imxrt1170", feature = "imxrt1180")))]
     MHz75,
     MHz80,
     MHz100,
@@ -108,6 +108,7 @@ pub enum SerialClockFrequency {
         feature = "imxrt1040",
         feature = "imxrt1060",
         feature = "imxrt1064",
+        feature = "imxrt1160",
         feature = "imxrt1170",
         feature = "imxrt1180"
     ))]