+++ CHANGELOG.md
@@ -10,6 +10,8 @@ Add support for new MCUS:
 - 1040
 - 1160
 
+Expose configuration commands in the common FlexSPI configuration block.
+
 ## [0.3.3] - 2024-10-26
 
 Add support for the 1180 family